backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
openai
httpx
stripe
paypalrestsdk
sqlalchemy
pymysql
passlib[bcrypt]
python-jose[cryptography]
pydantic[email]